// packages/core/src/lib/reaction/enums/event.enum.ts
export enum ReactionEvent {
  SET_GUESSES = 'SET_GUESSES',
  SET_DURATION = 'SET_DURATION',
  SET_ID = 'SET_ID',
  SET_GUESS_STATUS = 'SET_GUESS_STATUS',
  SET_REACTION_STATUS = 'SET_REACTION_STATUS',
  SET_IS_GUESSED = 'SET_IS_GUESSED',
  SET_STARTED_AT = 'SET_STARTED_AT',
  SET_COMPLETED_AT = 'SET_COMPLETED_AT',

  EVENT_COMMENCE_COUNTDOWN = 'EVENT_COMMENCE_COUNTDOWN',
  EVENT_START_ANIMATION = 'EVENT_START_ANIMATION',
  EVENT_STOP_ANIMATION = 'EVENT_STOP_ANIMATION',
  EVENT_ADD_GUESS = 'EVENT_ADD_GUESS',
  EVENT_COMPLETED = 'EVENT_COMPLETED',
  EVENT_FAILED = 'EVENT_FAILED',
}